requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-telegram-bot==20.7
pyjwt==2.8.0
python-dotenv==1.0.0
werkzeug==3.0.1
requests==2.31.0
aiohttp==3.9.1
